correcting type

+++ frameworks/zix-http3/src/main.zig
@@ -38,16 +38,16 @@ const TLS_KEY_DEFAULT: []const u8 = "/etc/zix-h3/server.key";
 const MAX_DATA_CHUNK: usize = 16 * 1024;
 
 // IP Maximum Transmission Unit
-const IP_MTU: usize = 1500;
+const IP_MTU: u64 = 1500;
 // Header IP, IPv4 use 20 bytes and IPv6 use 40 bytes
-const HEADER_IP: usize = 40; 
+const HEADER_IP: u64 = 40; 
 // UDP Header from source port, destination port, len of udp header + payload, checksum
-const UDP_HEADER: usize = 8;
+const UDP_HEADER: u64 = 8;
 // Max datagram came from value of IP math on 1500 MTU (IPv4 & IPv6 headers alloc)
 // best value is 1452 since involve two models IP. But counting use IPv6 instead IPv4
 // best_option = IP_MTU - HEADER_IP - UDP_HEADER
 // the default is 1200 as safe floor.
-const MAX_DATAGRAM_SIZE: usize = IP_MTU - HEADER_IP - UDP_HEADER;
+const MAX_DATAGRAM_SIZE: u64 = IP_MTU - HEADER_IP - UDP_HEADER;
 
 // Static directory, overridable via ARENA_DATA (default /data, the container mount point).
 var g_static_base: []const u8 = "/data/static/";
